add magic mirror scripts
netv2-status is a helper script that reflects serial port
data to port 6502 on the web server so magic mirror can scrape
netv2 status. netv2-status should be started as a daemon,
on the default system it's done using pm2.

netv2-term is the magic mirror module that does the rendering
of the reflected status data. netv2-term should reside in
the MagicMirror/modules directory.